requirements.txt
playwright
python-dotenv
python-telegram-bot
schedule


